feat(marks): refactor and keep view information
** Refactor

Previously most functions used to "get" a mark returned a position,
changed the line number and sometimes changed even the current buffer.

Now functions return a {x}fmark_T making calling context aware whether
the mark is in another buffer without arcane casting. A new function is
provided for switching to the mark buffer and returning a flag style
Enum to convey what happen in the movement. If the cursor changed, line,
columns, if it changed buffer, etc.

The function to get named mark was split into multiple functions.

- mark_get() -> fmark_T
- mark_get_global() -> xfmark_T
- mark_get_local() -> fmark_T
  - mark_get_motion() -> fmark_T
  - mark_get_visual() -> fmark_T

Functions that manage the changelist and jumplist were also modified to
return mark types.

- get_jumplist -> fmark_T
- gets_changelist -> fmark_T

This refactor is also seen mainly on normal.c, where all the mark
movement has been siphoned through one function nv_gomark, while the
other functions handle getting the mark and setting their movement
flags. To handle whether context marks should be left, etc.

** Mark View

While doing the refactor the concept of a mark view was also
implemented:

The view of a mark is the number of lines between the mark position on
creation and the window topline. This allows for moving not only back to
the position of a mark but having the window look similar to when the
mark was defined. This is done by carrying and extra element in the
fmark_T struct.

*** User space features

1. There's a new option. jumpoptions+=view enables the mark view restoring
automatically when using the jumplist and changelist. ^O ^I g; g,

2. There's new mappings to move to a mark and set its view.
   [count]' , [count]` , z' , z`

** Limitations

- The view information is not saved in shada.
- Calls to get_mark should copy the value in the pointer since we are
  using pos_to_mark() to wrap and provide a homogeneous interfaces. This
  was also a limitation in the previous state of things.

Co-authored-by: Gregory Anders <8965202+gpanders@users.noreply.github.com>
Co-authored-by: James McCoy <jamessan@jamessan.com>